Add files via upload
This is an R markdown file to serve as the skeleton for our report. It includes Shantel's data loading and cleaning code, as well as her PCA code. I've included my section on classification modeling as well. I've changed some of the styling of comments and spacing to improve readability (from my perspective). I also implemented a training and testing split under the data cleaning code. In the PCA code, I made it to where only the training set is used, so the prcomp() function hasn't seen the testing data when it is time to transform the model for classification modeling.